+++ packages/thirdweb/src/react/core/design-system/CustomThemeProvider.tsx
@@ -22,7 +22,7 @@ export function CustomThemeProvider(props: {
 }
 
 export function parseTheme(theme: "light" | "dark" | Theme | undefined): Theme {
-  if (!theme) {
+  if (!theme || !isValidTheme(theme)) {
     return darkThemeObj;
   }
 
@@ -35,6 +35,14 @@ export function parseTheme(theme: "light" | "dark" | Theme | undefined): Theme {
   }
 
   return themeObj;
+}
+
+export function isValidTheme(theme: unknown): theme is Theme {
+  return (
+    theme === "dark" ||
+    theme === "light" ||
+    (typeof theme === "object" && theme !== null && "colors" in theme)
+  );
 }
 
 /**