+++ packages/thirdweb/src/react/web/ui/components/buttons.tsx
@@ -1,5 +1,9 @@
 "use client";
-import { useCustomTheme } from "../../../core/design-system/CustomThemeProvider.js";
+import {
+  isValidTheme,
+  parseTheme,
+  useCustomTheme,
+} from "../../../core/design-system/CustomThemeProvider.js";
 import {
   type Theme,
   fontSize,
@@ -17,7 +21,8 @@ export type ButtonProps = {
 };
 
 export const Button = /* @__PURE__ */ StyledButton((props: ButtonProps) => {
-  const theme = useCustomTheme();
+  const _theme = useCustomTheme();
+  const theme = isValidTheme(props.theme) ? parseTheme(props.theme) : _theme;
   if (props.unstyled) {
     return {};
   }