+++ legacy_packages/auth/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @thirdweb-dev/auth
 
+## 4.1.97
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/wallets@2.5.39
+
+## 4.1.96
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/wallets@2.5.38
+
+## 4.1.95
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/wallets@2.5.37
+
+## 4.1.94
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/wallets@2.5.36
+
 ## 4.1.93
 
 ### Patch Changes